+++ HISTORY.md
@@ -1,3 +1,8 @@
+- plotter_enhanced:
+
+     - From Masoud: New transform mask_expander (masks out weights, relative to a threshold
+       value in 2D, in 1D along the time axis, or in 1D along the frequency axis)
+
 - Version 8:
 
      - Backwards-incompatible: jsoncpp library now required (see README.md for installation instructions)
@@ -46,19 +51,19 @@
        easier to interpret, since they can be generated in 1-1 correspondence with intensity
        waterfall plots.  (See example3_chime for an example)
 
-- Version 4
+- Version 4:
 
   - plotter_transform: fix bug which sometimes caused outlier intensities to be superfluously masked
   
   - plotter_transform: determine color mapping using outlier-clipped mean/variance (without this, the
     plots were sometimes misleading!)
 
-- Version 3
+- Version 3:
 
   - From Masoud: New transform badchannel_mask, which reads a text file containing a list of bad 
     frequency ranges, and masks channels which overlap (currently python-only)
 
-- Version 2
+- Version 2:
 
   - new transform chime_file_writer, which writes a stream to disk in CHIME hdf5 format.